functions to generate dataset ready

+++ generate.sh
@@ -5,10 +5,23 @@ if ! [[ "$nb" =~ ^[0-9]+$ ]]
 then
 	echo "Input passed not a number"
 else
+	echo "begin panorama cration" > log.txt
 	for sample in $(ls data)
 	do
-		echo ${sample}
-		ls -d data/$sample/ground_truth/* | head -n "$nb" | xargs ./image-stitching
-		mv out.jpg data/$sample/blurred/
+		if [ -e "data/$sample/blurred/out.jpg" ]
+		then
+			echo "$sample skipped"
+			echo "$sample skipped" >> log.txt
+			rm data/$sample/blurred/out.jpg
+		else
+			echo "$sample"
+			ls -d data/$sample/ground_truth/* | head -n "$nb" | xargs ./image-stitching
+			if [ -e "out.jpg" ]
+			then
+				mv out.jpg data/$sample/blurred/
+			else
+				echo "$sample panorama failed" >> log.txt
+			fi
+		fi
 	done
 fi
